+++ src/components/Common/Header/CartModal/CartModal.scss
@@ -7,9 +7,14 @@
   width: 100%;
   height: 514px;
   background-color: $primary-white;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-  display: flex;
+  box-shadow: 2px 4px 10px 0px #AFAFAF40;
   flex-direction: column;
+  margin-top: 17px;
+  @include xl {
+    max-width: 497px;
+    height: 586px;
+
+  }
 
   &__header {
     padding: 16px 20px 0;
@@ -44,11 +49,17 @@
     align-items: center;
     justify-content: center;
     height: 100%;
+    
+    @include xl {
+      justify-content: flex-start;
+      padding-top: 225px; 
+    }
   }
+  
 
   &__image-wrapper {
     @include icon-size(49px);
-    margin-bottom: 16px;
+    margin-bottom: 20px;
   }
 
   &__image {
@@ -61,10 +72,32 @@
     line-height: 22px;
     letter-spacing: 1.97px;
     text-align: center;
-    padding-left: 80px;
-    padding-right: 80px;
+    width: 145px;
     
   }
+
+    &__history {
+      display: none;
+
+      @include xl {
+        display: block;
+        height: 48px;
+        width: 206;
+        background-color: transparent;
+        @include bold-text;
+        line-height: 20px;
+        letter-spacing: 1.97px;
+        padding: 14px 15px;
+        margin-top: 135px;
+        margin-bottom: 44px;
+        border-radius: 1px;
+       
+
+      }
+
+
+
+    }
 
   &__item {
     display: flex;
